refactor: Update import paths for components in page.jsx
