+++ app/page.jsx
@@ -1,118 +1,15 @@
 'use client'
-import { animated, useSpringRef, useTransition } from 'react-spring'
-import { useLayoutEffect, useState } from 'react'
-import Techs from 'app/components/Techs'
-import NextImage from 'next/image'
-import AnimatedLink from 'app/components/AnimatedLink'
 
-const IMAGES = [
-  'https://fiverr-res.cloudinary.com/images/t_main1,q_auto,f_auto,q_auto,f_auto/gigs/154443660/original/734570cec0de955789ff0acd80caad3d582b85e0/create-a-generative-art-piece-for-you.png',
-  'https://images.squarespace-cdn.com/content/v1/5c77350965a707ed1710a1bc/1592330659753-70M66LGEPXFTQ8S716MX/Generative+Art+by+Mark+Stock+-+Gyre+35700.jpg',
-  'https://cdn.pixabay.com/photo/2018/09/04/09/12/generative-art-3653275_1280.jpg',
-]
+import Techs from 'app/components/Techs'
+import About from '@components/About'
+import DemoAnimation from '@components/DemoAnimation'
 
 export default function Home() {
-  const [activeIndex, setActiveIndex] = useState(0)
-  const springApi = useSpringRef()
-
-  const transitions = useTransition(activeIndex, {
-    from: {
-      clipPath: 'polygon(0% 0%, 0% 100%, 0% 100%, 0% 0%)',
-    },
-    enter: {
-      clipPath: 'polygon(0% 0%, 0% 100%, 100% 100%, 100% 0%)',
-    },
-    leave: {
-      clipPath: 'polygon(100% 0%, 100% 100%, 100% 100%, 100% 0%)',
-    },
-    onRest: (_springs, _ctrl, item) => {
-      if (activeIndex === item)
-        setActiveIndex(activeIndex === IMAGES.length - 1 ? 0 : activeIndex + 1)
-    },
-    exitBeforeEnter: true,
-    config: {
-      duration: 4000,
-    },
-    delay: 1000,
-    ref: springApi,
-  })
-
-  useLayoutEffect(() => {
-    springApi.start()
-  }, [activeIndex, springApi])
-
   return (
     <>
-      <div className="mb-5">
-        Hi 👋 I’m Gökhan, a software engineer at{' '}
-        <AnimatedLink
-          href="https://teknasyon.com/en/"
-          color='teknasyon'
-        >
-          Teknasyon
-        </AnimatedLink>
-        . I&rsquo;m working on{' '}
-        <AnimatedLink
-          href="https://rockads.com/"
-          color="rockads"
-        >
-          rockads.com
-        </AnimatedLink>{' '}
-        and B2B projects & their open-source projects. I love open source
-        development and I build things on my{' '}
-        <AnimatedLink
-          href="https://github.com/gokh4nozturk"
-        >
-          GitHub profile.
-        </AnimatedLink>{' '}
-        {/* Example projects:{' '}
-        <a
-          href="https://github.com/Teknasyon/rocket-ui"
-          target="_blank"
-          className="decoration-[#4C5BE1] underline-offset-4 hover:text-[#4C5BE1] max-sm:underline md:hover:underline"
-        >
-          Rocket UI
-        </a>
-        , I love coding, reading and taking photos. */}
-        <br />
-        You can follow me on{' '}
-        <AnimatedLink
-          href="https://twitter.com/gokh4nozturk"
-        >
-          Twitter
-        </AnimatedLink>
-        {'  '}&{' '}
-        <AnimatedLink
-          href="https://www.linkedin.com/in/gokhannozturk/"
-          target="_blank"
-        >
-          LinkedIn
-        </AnimatedLink>
-        {/* , or reach out via{' '}
-        <a
-          href="mailto:me@gokhanozturk.io?subject=Mail from your Website"
-          className="underline-offset-4 max-sm:underline md:hover:underline"
-        >
-          Email.
-        </a>{' '} */}
-      </div>
-      <div className="w-full">
-        technologies I use: <Techs />
-      </div>
-      <div className="mt-10">
-        {transitions((springs, item) => (
-          <animated.div style={springs}>
-            <NextImage
-              className="rounded-sm"
-              src={IMAGES[item]}
-              priority
-              width={800}
-              height={600}
-              alt="Gökhan Öztürk"
-            />
-          </animated.div>
-        ))}
-      </div>
+      <About />
+      <Techs />
+      <DemoAnimation />
     </>
   )
 }
